refactor!: update the type of total count from uint32 to int64
BREAKING CHANGE: The TotalCount field of BaseWithTotalCountResponse DTO will be changed from uint32 to int64

Since postgresql is the default persistence layer and the count() func returned from postgresql will be
bigint(https://www.postgresql.org/docs/17/functions-aggregate.html) with value range int64, we should
update the total count value type from uint32 to int64 to avoid integer overflow issue, where the max
of uint32 is 4,294,967,295 and the max of int64 is 9,223,372,036,854,775,807.

+++ dtos/common/base_test.go
@@ -56,7 +56,7 @@ func TestNewBaseWithTotalCountResponse(t *testing.T) {
 	expectedRequestId := "123456"
 	expectedStatusCode := 200
 	expectedMessage := "unit test message"
-	expectedTotalCount := uint32(3)
+	expectedTotalCount := int64(3)
 	actual := NewBaseWithTotalCountResponse(expectedRequestId, expectedMessage, expectedStatusCode, expectedTotalCount)
 
 	assert.Equal(t, expectedRequestId, actual.RequestId)
